Revert "start aconfig storage daemon"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2def2bc9-4177-4451-930d-96612adf7d95). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I876df31f5bb3c6b070ed60886ed40268e79a3424