Revert "Declare init.*.rc modules in Android.bp"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/25b72c41-5358-48e3-bfee-b89891cc891a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4d46033b4668d85965c8d573f0a9f58b3ed98322