Revert "Declare init.*.rc modules in Android.bp"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9dc4078f-b03d-4863-a37c-bf8335056a4f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3ac6190d02f64761920a52acb77d635767092f43