Revert "turn on aconfigd in init"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0596ae1f-5830-45f1-b0d4-f059d23f0dcd). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic922af68a288b3b7b4e00b8f03887c1d9efa65d5