Revert "turn on aconfigd in init"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0cc45563-6ba6-4ce8-ac50-f72364fe0085). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0f92a03408ce1335d9d7648e0785f7b11868834b