Revert "turn on aconfigd in init"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1878a190-1b5c-43d7-abfd-9cc03bca5b27). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Idfce16fc8b030ee05477a9d8a783fc1eccf91443