Revert "turn on aconfigd in init"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/71e228dd-bab1-4751-98fe-ecfc0f56f796). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iff51138986fd69890532908b90be2b23a9bf1513