Revert "Build init-debug.rc with Soong"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e80426ec-8cbb-49fd-a711-20e9a2fc0649). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1ffb420df69eb38cc2bba26ec93592cdc4845fbb