Revert "storageproxyd: Add arguments for storage mapping and max..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7e14ae78-8e5e-498f-8574-30d6cbd48d31). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I75865d100cbd6b5d8a523dcbedb72da9e9b45de8