Create app-metadata directory
App metadata bundles, also known as Android Safety Labels (ASL), contains
information about the app's privacy and security practices. This
information is used to help users make more informed choices when, for
example, granting access to permissions.

ASL can currently only be preloaded on the system image or distributed by
the installer and written to a file in the app's codePath. To support
embedded ASL in APK we need to extract ASLs from APKs to a writeable
location. For non-preloaded apps we can write to the app's codePath like
the installer provided ASLs, however, we need to create a new writable
directory for preloaded apps located on read only partitions.

Bug: 336618214
Test: manual
Change-Id: I651b2dab45c3132d8467c507dc4ee304001f73f5