Add support to restore selinux context of /data
For ARC devices, it's possible for the selinux context of /data
to be corrupted, which results in ARC not booting. To combat
this, an init command was added to check the selinux context of
select subdirectories within /data. If the context does not match
what is expected, restorecon is run on /data. Support was also
added for additional selinux android restorecon flags that are
needed. Lastly, boot properties are set that will later be picked
up by ArcMetricsService to record status and duration of this
operation.

Test: tast run dut arc.RestoreData.whole_data_vm
Test: tast run dut arc.RestoreData.app_data_vm
Bug: 210155681

Change-Id: I9d5b64c48d21e83fb8f03dac261efc530c5f8aa6