Introduce ARCVM-specific init_first_stage and init_second_stage
These variants compile with the -DARCVM=1 flag, which will allow us
to include ARC-specific logic in early init. They can be included
as-is to the ARC PRODUCT_PACKAGES because they override the existing
first and second-stage init binaries when installed.

This change also creates a new set of Android-only init dependencies
to be shared between standard Android and ARCVM.

Bug: 325538592
Test: build bertha_x86_64 and bertha_arm64
Change-Id: Id733049f65f12eb5cce14d7bfc345e199c5c6d4e