Revert "Set ro.fstype.data when /data is mounted"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/14e07f88-0bb2-4ffd-a9c1-465f617ae4cf). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 343561082
Change-Id: I69cfa45ae07af680e09ceaeb305cb87992baeb6d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com