Revert "Set ro.fstype.data when /data is mounted"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4ac43f6d-5869-4f87-bbcd-4f193a0ee080). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 343561216
Change-Id: I1be33ecbf84d25fefa259b05da5033968f237488
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com