Revert "Set ro.fstype.data when /data is mounted"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d0b217a3-caff-4961-b711-3719377e7b41). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 343560782
Change-Id: I299e5f5239d85003546aeef8765b0fc6bfd57b09
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com