Revert "Set ro.fstype.data when /data is mounted"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/799bdb33-8893-43c9-93b7-788f88e1f1ed). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 343560933
Change-Id: Ia2485d8411dc626bdc2b96192c14c2847e97457f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com