Revert "Set ro.fstype.data when /data is mounted"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0b124fb7-cef4-4e8f-9c07-4f926acb81c4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 343560982
Change-Id: I4ca63e4b85c6fd20fbc1ff0f4b8aaf7d1f39bb01
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com