Revert "Set ro.fstype.data when /data is mounted"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/03f3df84-5ca4-48b9-9fb7-3b4cda663aea). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 343560867
Change-Id: Ie0535ebed984aa01563d8d9cfb6ec4ac44b1e9ca
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com