Revert "Set ro.fstype.data when /data is mounted"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1e7b0974-751a-45ed-8d0b-37c4244531e3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 343560724
Change-Id: Ia4e9717bb9ec4c043a0e6ffd72782a3b90ed1951
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com