Revert "Set ro.fstype.data when /data is mounted"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/64e13983-6cff-4737-b49b-813e6298e430). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 343560979
Change-Id: If6044c10db55ba84107190523ca45338bd922a7a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com