Revert "Revert "Make foreground and background I/O priority diff..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/172a75c4-2eae-4027-b555-de04f31d0073). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1c4a89293e2d2256e3c95c52b1fe889bb408b436
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com