Revert "Revert "Make foreground and background I/O priority diff..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2d274a56-9b2a-45f0-bbb8-4d23885ac3b9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I06ca44ddf60a0c757f9e037aeaec7ff51205f3b0
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com