Revert "reupload: libsnapshot: set thread priority"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f389596f-5323-4de7-82bc-6be5038b154d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 348046367
Change-Id: I9c96c413c628d8615e612868e98bd453c89597ef
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com