Disable images in recovery
If a previous snapshot was not cleaned up, we want to mark the snapshots
for cleanup upon the next reboot (in recovery mode). Since we cannot
mount /data in recovery, attempting to delete the snapshots alone will
return an error.

Bug: 340721205
Test: apply OTA -> reboot recovery -> install sideload OTA (confirm
success) -> reboot and
confirm that no snapshots exist under /data/gsi/ota

Change-Id: I2894f5fbc1dd534510bc2b6581f54e53133602d7