Revert "Rewrite `init_vendor` using select syntax"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6ca31272-2a09-4dd0-aabd-968227dbf7a5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3c8d890325c68efbff86e01724c644783670251a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com