Use MNT_DETACH in unmounting tests
Aconfig flagging is migrating to a new storage backend that makes heavy
use of the /metadata partition. It is causing failures in tests like
vts_libsnapshot_test, which unmount and remount this partition. The
failed unmounts are due to the partition being busy. Switching the
unmount flags to MNT_DETACH resolves the failures, which from manpages,
makes the partition unavaiable for new requests, then lazily unmounts.

Bug: 349807180
Test: m vts_libsnapshot_test
Change-Id: Ie4f8cadcad189b13b9eba1eb15c5251002e8138a