fs_mgr: check available space before creating zram backing device
This CL adds a check to ensure that there is enough free space on
the device before creating the zram backing device.
If there is not enough space, the zram backing device will not be
created.

Adding a threshold 512 MB to prevent the zram backing device from
running out of all of the data space.

Bug: 322821637
Bug: 326877642
Test: Fill up userdata, enlarge zram backing device size on next
boot.
Change-Id: Iecb60bb90a9e2cc9ae53c824449c0ca222a11f76

Change-Id: I135b4beee2a8a21f8214f6579368a8d0d8792255