Revert "storageproxy: set a property when secure storage becomes rw"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/06318941-ad17-4f11-ad6a-6d89fb6eee89). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6a088f11d4fcf7736adcf7eae57ef682e4365ddf
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com