Revert "charger: fix show qustion mark when start to draw UI"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/44437649-e85a-47d0-9171-66a76fda0592). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I948e68dea79a3bb3d1128880938d7f0282d2ec09
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com