Revert "Add trusty_rkp_set_uds_cert for UdsCert provisioning"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4bceb103-d085-44d5-9993-5c5b9cd9e70b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4154f2e1711bba38985fc674c4ed2e83b9c464ae
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com