Revert "Migrate init.environ.rc module to Soong"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/51ebb3ca-2cf1-40eb-8f5a-226ae52a988e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 356587212
Change-Id: I509470de51fe1dc30ca59f54f25dda078ecb6473
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com