Revert "trusty: keymaster: Add commandline option to specify dev..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/43102b04-da0f-4a1d-8a7b-bdd0e49d9832). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iba428b822e5a351a75d94475537eb18f1511d6be
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com