Revert "trusty: gatekeeper: Add device option"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bf8361e3-5465-46a6-893d-4f983f8c13f8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id19b63824740fbd66d2610a0665b7096f1687170
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com