WIP: port adb-remount-test.sh to Python.
This script is really difficult to read, and even more difficult to
debug. It's not easy to correlate failures with actual points in the
test script. And it's difficult to add debugging logic because all the
stdout is mingled together.

This Python rewrite tries to be as faithful to the original as possible,
with some simplifications (eg all the usb/slot checks are gone).

Bug: N/A
Test: adb-remount-test.py
Change-Id: I7a0f124cec1fdfd34ce17409d63d8ab6eabf58d3