Revert "trusty: libtrusty: Add vsock support"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/42a154f7-dc4a-440f-80e3-fd03ceec319b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Icd4336537c06c35ee979addaaccdd6f37591d8f7
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com