Switch to linker allocator for all allocations.
There is more than one function that can allocate in the fallback
path. Therefore, make sure that all functions that can allocate have
switched to the linker allocator before allocating.

Add a ScopedUseLinkerAllocator class to do the switching to make it
easier to have the same code everywhere.

Add two tests to verify that no allocations are occuring during
a tombstone or backtrace on the fallback path.

Remove the debuggerd_fallback_tombstone() function since it only
enables the linker callback and then calls engrave_tombstone_ucontext.
The enabling is now done with the ScopedUseLinkerAllocator.

Bug: 359692763

Test: Ran debuggerd -b <PID> and debuggerd <PID> on a process that
Test: goes throught the fallback path.
Test: Unit tests pass along with two new tests.
Test: Forced a fallback process to crash and verified tombstone generated.
Change-Id: I2710921076634eac97f41bec8c3a29c1d75ae5ec