Revert "libsnapshot: Move snapshot metadata to super partition."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/64e3fdef-eff5-4f6a-aea0-30e1e836d227). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5506da308f1aa687d52ace2ec60eaa85e9845bbc
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com