Revert "Skip F2FS formatting for dev option enabled devices"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0b47782a-7c56-4831-a1c2-0f84f2b4bfb7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibaf66da72842a110db10974f7742860c916a4f43
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com