Revert "init.rc: Fix entropy pool initialization"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/447cbe77-0a8c-4af8-b55e-c4131129383f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie417f6b6afd05e7a1c2dd201630db17dd4933096
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com