Revert "Skip F2FS formatting for dev option enabled devices"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/765d4a4e-99ae-4f40-a1ee-2f73442bf2ac). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6965cf992f3fc66dd4f3a9ff2af26abd8e0f4759
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com