Revert "Skip F2FS formatting for dev option enabled devices"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/260168de-c591-482d-aabe-e8d51c839b33). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1416e2f40a8b4a1247c22ada84ffe6f4223f34a1
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com