Revert "Skip F2FS formatting for dev option enabled devices"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/df057651-971e-41a1-baf7-2f081a19e377). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8fbbd3895a888edd600c2604a2adad72e9c0c486
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com