Revert "Skip F2FS formatting for dev option enabled devices"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ae48f71b-5b0f-427c-a936-5f27483aafb9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I05872b011b5230f9532b28d923b5d487f79c5ade
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com