Revert "Log lock time if it takes too long"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/11d211a3-4a07-4c5a-a307-28752c246507). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I91addbcbda694ef761a2597ea631b81df0a199f6
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com