Add libprocessgroup_vendor
The toybox_vendor module used to bring in libprocessgroup to the vendor
image, but it was updated to remove that dependency. Vendor tests still
require the library, and while these tests build succesfully they fail
to run because the aosp_cf_x86_64 vendor images no longer have any
modules that depend on the vendor libprocessgroup, so libprocessgroup is
no longer included on the vendor images. Add them back.

Test: atest liblyric.api_test -- --abi x86_64
Bug: 364392168
Bug: 349105928
Change-Id: I5e3eea4c7812f6920818f660a3fba1876290c39a