init: Ensure fsck tools are available in first-stage init.
Due to quirks in how the first-stage ramdisk is packaged, e2fsck is not
normally available. This prevents check_fs() from working if the "needs
fsck" bit is set on the partition.

This change ensures that the fsck binaries are copied just as snapuserd
is.

It's possible this change will cause some kind of boot time regression
on devices with the "check" fs_mgr flag. If so, the proper fix is to
remove the "check" flag from metadata's fstab entry. This flag has been
non-functional since at least init_boot was introduced (Android T).
Removing it will stop preventative fsck from running, but normal fsck
will still work.

Test: logcat shows fsck running on /metadata
Bug: 364764252
Change-Id: Ibb150263de605c8ddc13e8533b5b835857421147