Revert "trusty: Add host build for rpmb_dev"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4a7c207a-2c43-4685-8842-972aed11c8f2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 365632893
Change-Id: I56b7ad142d89d7375bafba643f2a587487ea259e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com