Revert "Remind the reader that they'll need to modify CTS too."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9b6a7c09-3eb4-4ec8-a92b-6aef020131db). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I33043ae1e9ae54f61f594af0dcf3e295891b99c9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com