Fix malformed kernel panic subreason
Summary: The kernel panic subreason has some aggressive logic that is tightly coupled with the assumption that bit errors are common in pstore logs.
The way subreasons are parsed are as follows:

1) Does the panic string match fully an expected value
-> if yes, return sysrq

2) Does the panic string subreason match a known subreason
-> if yes, return a mapped subreason

3) Crawl through the subreason and fix assumed bit errors and return

The bit error logic is something like this:
1) Provide a known string token (needle) that could be in the panic str
2) Compute the max bit error value based on string length
3) Walk through the panic string one character at a time
4) If we find a substring that matches within bit error, replace that substring no matter what

The bit error is calculated as:
1) Allow up to 2 bits (kBitErrorRate = 8 ) per character
2) Allow [n * 2] bits per string up to 2 characters
3) Allow [(n * 2) - 1] bits per string above 2 characters (not exactly the algorithm but close enough for this specific case)

Walking through our first case, we are presented with panicstr:
```
sysrq_triggered_crash
```

The list of known strings contains warm. We will walk through panicstr
one character at a time. When we hit gere we check potential bit
errors.

gere ~= warm
g (01100111)
w (01010111)
 00110000
e (01100101)
a (01100001)
 00000100
r (01010010)
r (01010010)
 00000000
e (01100101)
m (01001101)
 00101000

Total differences 5 bits where we never exceed the bit error threshold
of the entire string.

There are probably ways to fix the fuzzy search or to walk through the
known list once again to see if the fuzzy search returns a known total
string, then if not use the existing string as is. This fix seemed the
most straightforward to add reasons that we expect to skip fuzzy search.
We could improve this by creating a HAL or something that returns
known reasons for vendor and OEM to bootstat.


Test: Successful Build on master branch

Change-Id: Icffe8d5cdbd7d86e2383a7f50e3c93f0ab215767
Signed-off-by: Abdelrahman Daim <adaim@meta.com>