Fix malformed kernel panic subreason
Summary:

The kernel panic subreason logic is tightly coupled with the assumption that bit errors are common in logs. The way subreasons are parsed is as follows:

1. Does the panic string match fully an expected value? If yes, return the matched value.
2. Does the panic string subreason match a known subreason? If yes, return the mapped subreason.
3. Crawl through the subreason and fix assumed bit errors and return.

The bit error logic works as follows:

1. Provide a known string token that could be in the panic string.
2. Compute the max bit error value based on string length.
3. Walk through the panic string one character at a time.
4. If a substring is found that matches within the bit error, replace that substring.

The bit error is calculated as:

1. Allow up to 2 bits per character.
2. Allow [n * 2] bits per string up to 2 characters.
3. Allow [(n * 2) - 1] bits per string above 2 characters.

Walking through a case, we have a panic string: `sysrq_triggered_crash`. The list of known strings contains `warm`. We walk through the panic string one character at a time. When we hit `gere`, we check potential bit errors.

`gere` is similar to `warm`, with a total difference of 5 bits, which is within the bit error threshold.

There are probably ways to improve the fuzzy search or to walk through the known list again to see if the fuzzy search returns a known total string. This fix seemed the most straightforward to add expected reasons and skip fuzzy search. We could improve this by creating a way to return known reasons for vendors.

We can add the reasons we see in logs.

Test: Successful Build on master branch

Change-Id: Icffe8d5cdbd7d86e2383a7f50e3c93f0ab215767
Signed-off-by: Abdelrahman Daim <adaim@meta.com>