healthd: never keep charger screen on longer than 60s
Summary: The healthd charger app is programmed to turn the screen
on briefly whenever the charging status changes, i.e. when
the charging cable is plugged or unplugged.
Some flaky charger/cable combinations may cause the
service to register the charger as quickly connecting and
disconnecting. This causes the screen to stay on indefinitely
causing the charging icon to burn in and battery to discharge.
Add a global timeout for how long the charger screen can stay
on, regardless of connected/disconnected status.

Test: Find a crappy charger that causes the charging to
connect and disconnect constantly.
Failing that, get a USB hub that has a non/off switch for each
connector that you can toggle about once a second.
As a last resort you can just continually unplug and plug the
usb cord quickly.
Observe the screen stays on while the device the charger
is connected/disconnected.
After 60s the screen should turn off, even if the charger
is being disconnected/re-connected.
It should only come back on after pressing the power button.

Change-Id: Ide2a6caea316991d63f062fbef11b297c3b5aa02
Signed-off-by: Abdelrahman Daim <adaim@meta.com>