Revert "init: Look for super partition only on a boot device"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f939c10a-05ba-4114-83f6-73c0ffde1c07). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 371150763
Change-Id: If00497bc73c1338643ee327c385eb5c44eec84b1
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com