Revert "init: Look for super partition only on a boot device"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8a2b9113-6f2d-4dbb-b78b-f8ee3d163b1d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 371510462
Bug: 371513237
Change-Id: I4508d50f297711f8cb6b117eedb36fafa9067c4c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com