Revert "init/epoll: clean up reorder-init-list warning"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/478d612b-96a6-447a-a6d0-5f871c81b004). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I467b195cc0c877483f2f90bb87cd2f5df310e728
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com