Revert "init: Look for super partition only on a boot device"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c8eb5dce-a4b1-4394-8970-8723299e67f0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4e757813890a86f89868bf1e627e9250733d18b8
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com