Revert "init: Look for super partition only on a boot device"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f5fe82da-b26f-44fa-9ccf-cf5826df86d0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9d4e107a75ced6788b8bfbb11b80134bbe70cb66
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com