Revert "init: Look for super partition only on a boot device"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d068d936-d02c-4a6c-b2a9-0504ace7056a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iea4c83081825399b17eb426c5067fe299d164a5a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com