remount: Make scratch filesystem with "nodiscard" flag
According to "man mkfs.ext4", the "discard" flag is useful for "SSD and
sparse / thin-provisioned storage".
For remount/overlayfs we use IImageManager to allocate and manage the
data blocks so the "discard" option don't benefit us.
Note that we already pass "nodiscard" when mounting the scratch
filesystem.
As a result, also explicitly pass nodiscard when calling mkfs.

Bug: 359464822
Test: adb-remount-test
Change-Id: Ia93d4e5014969fb6a976a6fcc15d912c32abf357