remount: Make scratch filesystem with "nodiscard" flag
According to "man mkfs.ext4", the "discard" flag is useful for "SSD and
sparse / thin-provisioned storage".
For remount/overlayfs we use IImageManager to allocate and manage the
data blocks so the "discard" option don't benefit us.
If the scratch device image is not mounted via dm-linear, but with
loopdev (direct file access), then the mkfs discard option could cause
file extents be deallocated and reallocated. This would break
IImageManager because IImageManager bookkeeps its allocated file block
extents.

Note that we already pass "nodiscard" when mounting the scratch
filesystem.
As a result, also explicitly pass nodiscard when calling mkfs.

Bug: 359464822
Change-Id: I57294d242700a923a31c84798d042a437ac57eba