Revert "init: Look for super partition only on a boot device"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/66897331-3329-4f94-9943-9787e0a21e1e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1d70cb40d5fc833c3c26676fbef46ec6db64c758
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com