init: Add the ability to find the boot device by partition UUID
The current mechanism for specifying boot devices on Android systems
involves passing a set of "bootdevices" though command line,
bootconfig, or device tree.

The bootdevices are specified as strings and, in general, need to
match a sysfs path but without the "/sys/devices" or
"/sys/devices/platform" prefix. The sysfs path is generally the path
to the closest parent of the block device that is a "platform" device.

As an example, if the sysfs path of the expected boot device is:
  /sys/devices/platform/soc@0/7c4000.mmc/mmc_host/mmc1/mmc1:0001/block/mmcblk1

The bootloader would specify it as "soc@0/7c4000.mmc" since:
* We strip off "/sys/devices/platform/"
* As we move up directories, we don't find one whose subsystem is
  "platform" until we get up to
  "/sys/devices/platform/soc@0/7c4000.mmc".

The current mechanism is a bit brittle. Specifically:
* The sysfs path isn't _really_ stable and can change across kernel
  upgrades. For instance, during one kernel upgrade the device tree
  for a product changed so that the root node changed from "soc" to
  "soc@0" and this changed all sysfs paths. In the past device tree
  folks have asserted that we shouldn't rely on dts node names to stay
  consistent, yet those node names are used to construct sysfs paths.
* For some devices, like USB, the path of the closest "platform"
  device tends to be the path of the USB controller. This means that
  if two USB disks are plugged in we can't guarantee which one will be
  identified as the boot device.

Add a new method of finding the boot device by passing the partition
UUID that we loaded the kernel from. Using the partition UUID to
identify the boot device is standard on Linux. You can see this
because when you're not using an initramfs you can use the syntax
"root=PARTUUID=<valid-uuid-id>[/PARTNROFF=n]" to specify the root.
Using the same idea for Android's boot code makes sense.

With this new method for finding the boot device, we can make the code
much more specific about matching sysfs paths. Once we find the sysfs
path for the kernel we can make sure that all of the other boot
partition are share the same "scsi" or "mmc" parent instead of going
all the way to the closest platform device.

There is the question of what we should do if the bootloader passes
_both_ an old style "bootdevices" and also a UUID. While we could
choose to simply ignore the old "bootdevices", we'll choose (for now)
to use the "bootdevices" to validate the device we found. We may want
to change this to just ignore the old "bootdevices" if a UUID is
specified or consider it an error to specify both. It's not really
necessary/a good idea for bootloaders to specify both because:
* As per above, the sysfs path is not guaranteed to be stable across
  kernel versions.
* It's not needed for security. While someone _could_ someone insert
  an external disk hacked to have the same UUID as the boot kernel,
  the worst they could do would be to cause the system not to boot. We
  already have verity in the system making sure our filesystems are
  not tampered with and this would also protect us from booting a
  tampered disk.

NOTE: this new method relies on the upstream kernel commit
74f4a8dc0dd8 ("block: add partition uuid into uevent as "PARTUUID"").

Bug: b:316324155
Test: Use UUID to boot

Change-Id: If824cb700ca3696a442a28e6ad02d7c522c3b495