Revert "init: Look for super partition only on a boot device"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a1493a86-25a4-4885-bda8-19336131a9ac). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic4cde2b294252ee42e8859e47b1f928594133f53
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com