Revert "Remove log spam."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/47115ec8-0119-46a2-b88f-294b0557546e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia48c038911b9481df205cc0612c68eb5457600ff
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com