Revert "Remove log spam."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e2999147-ee53-4551-85cb-2052dd7264b4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I27ea7f26ecf8cb92617be996f3be61753fc204e4
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com