Revert "Log lock time if it takes too long"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/92df1a41-5c6e-483a-a0cb-ea40291fc05a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I79816615518dc74e20d2ad9507d1a9fbb23f0fec
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com