Revert "Remove log spam."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9d89b961-7783-4d7e-93ad-6d089a58a83d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I78e607a50bf7335114caa487b816661593a333cc
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com