Revert "Remove log spam."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/529273c1-e5ac-4f9c-975f-381cde699d13). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0ee43b01524ba227e05a70b5e939e94eee52b2e3
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com