Revert "Remove log spam."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0392adaa-a0a8-4cde-abd3-cc5b3e0c8dc5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibc345d00c84958bdd82aea9fe1377e8c68e9dcf3
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com