Revert "Pin KeyMint dependency to correct/specific version"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/67fc686b-e7e5-4b94-ac43-dbd3973602cf). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3ad4979af0002671af73dbe25e54f0d7e1d1cd8f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com