Revert "support f2fs device aliasing feature"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/13e1aaa7-52aa-4c36-9ca3-0a1e951b618c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 376210371
Change-Id: I199a38ee2fe6cf50dd205886fb4001014731469f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com