Revert "Remove log spam."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ae66873a-4f6d-403c-bafa-9810c98f3d36). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8245d1a75e396e4c3077f5d2d4c497271d5f0ea9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com