Revert "libsnapshot: Fix MapAllSnapshotsWithoutSlotSwitch test"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7a78c6ce-9c69-4ea1-88c3-42421523f070). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I54f0c4fbf37e1499d68278e803290de845fd799e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com