Revert "Add proposed trendy teams for VTS modules"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/47bd450f-8b63-4d65-a670-138c99c944c3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Icb0e999672b370e96778da82255bc83113de2efd
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com