Revert "Set input thread priority to RT - try 3"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/45240238-6265-4b0d-ac60-c8e5671aca8c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 377739155
Change-Id: I2e431c78f69e3d9a1f3991cc670d6749219a46fa
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com