libprocessgroup: memcg default version 1 -> 2
The default memcg version is 1 in cgroups.json. Update the default to
version 2. Vendor overrides of cgroups.json [1] (to use v1 with the v2
default here) remain supported for kernels which still support memcg v1,
but core features of v1 have already been deprecated and removed, and it
is expected v1 will eventually be removed from the kernel entirely.

[1] https://source.android.com/docs/core/perf/cgroups#per-api-level-task-profiles

Test: (cuttlefish) adb shell grep memory /sys/fs/cgroup/cgroup.subtree_control
Bug: 377579705
Change-Id: I5b985fb4348b92d49022760331c0adf113bcdf04