Revert "Deprecating libvendorsupport_llndk_headers"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1276d1b9-3cf3-4869-8ec7-99432169883d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 378039242
Change-Id: I97bfa69265da568abff9abd52b3dd82f2205017d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com