Revert "Remove log spam."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ccaf2989-0107-4852-aa2c-d1b6b09ea1d5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5dc737413c5d9506f1bcaaaaca7469208a751a48
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com