Revert "Remove dependencies on the 1-variant fallback"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7a7afc1e-8502-46fe-b63c-d5098ead1c52). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib638a8b63efd201a6894780d1fddeb5c256ecdce
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com