Revert "Remove log spam."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/665ebe6f-28e9-416c-9159-95e414a73a62). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6b7f8d9f6a12ff46cc6cc32604e6922fdd6be3fd
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com