Revert "Remove log spam."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/819635ab-4b50-47b0-87a7-f1c86efa0cb3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I79062bbb5eba07096cc0b7cb741dd3b119be2c8f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com