Revert "Set the proper FEATURE_HARDWARE_KEYSTORE version"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3df9b93e-5a29-49c8-970f-7c44c604fc21). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iceb4991a17eea568ec39c5f4bafa670eb090ed7b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com